Lib.Formatting

Documentation

prettify :: Diary -> String

The prettify function converts debt diary into human readable text

>>> prettify diary
"Diary title
 diary note
 records:
 Idx: 0
   person: John Doe",
   note:
   amount: 14.14
   ts: 2011-11-19"

prettifyNth :: Int -> Diary -> String

The prettifyNth function converts nth debt record into human readable text

>>> prettifyNth idx diary
"Idx: 1
  person: John Doe
  note:
  amount: 22.22
  ts: 2022-11-1"

prettifyTable :: Diary -> String

The prettifyTable sum debt per user and converts into human readable table

>>> prettifyTable diary
"| Name     | Amount |
 | John Doe | 36.36  |"

prettifyBalance :: Diary -> String

The prettifyBalance sum all debt and create human readable string

>>> prettifyBalance diary
"Balance: 300.33"

newLines :: [String] -> String

The newLines function joins list of string with new line character.

>>> newLines ["a", "b", "c"]
"a\nb\nc"