Lib.Logics

Documentation

deleteRecord :: Int -> Diary -> Diary

The deleteRecord function deletes specified record by index from diary

>>> deleteRecord 0 diary

appendRecord :: Debt -> Diary -> Diary

The appendRecord function append debt record to diary

>>> appendRecord record diary

updateRecord :: Int -> Debt -> Diary -> Diary

The updateRecord function updates record in diary at specified index. If index does not exists, it will append record

>>> updateRecord 0 record diary

maxRecordIdx :: Diary -> Int

The maxRecordIdx function calculates maximal record index.

>>> maxRecordIdx diary
5