Lib.Operations

Documentation

parseDiary :: FilePath -> IO (Either Error Diary)

The parseDiary function reads file from specified file path and decodes Diary from it's content (in JSON format). Error handles invalid access to file or invalid format.

>>> parseDiary "debt-diary.json"
Diary "Fancy title" "note" []

apply :: (a -> IO (Either Error b)) -> a -> IO ()

The apply function takes function and value, passes value to function and checks result Whenever Left is encountered, it calls exitWithErrorMessage

apply2 :: Show a => (b -> IO (Either Error c)) -> Either a b -> IO ()

The apply2 function takes function and parameter, extracts Right from parameter and pass it's value to function. Checks result of function and returns Right value. Whenever Left is encountered, it calls exitWithErrorMessage

loadIndex :: String -> Int -> Either Error Int

The loadIndex function checks wheter value can be used as index Checks idx >= 0 && <= upperBound

saveDiary :: FilePath -> Diary -> IO (Either Error ())

The saveDiary function encodes diary into JSON and saves to filePath. Error handles invalid write access.

>>> saveDiary "debt-diary.json" diary

promptDebtRecord :: IO Debt

The promptDebtRecord function promts user for data and returns new Debt

promptDiary :: IO Diary

The promptDiary function promts user for data and returns new Diary

exitWithErrorMessage :: Show s => s -> IO ()

The exitWithErrorMessage function prints message to stderr and exit with status 2

>>> exitWithErrorMessage UnableReadFile